Extract getContainerAddress
