+++ discovery-aws-api-async/src/main/scala/akka/discovery/awsapi/ecs/AsyncEcsDiscovery.scala
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2017-2019 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.discovery.awsapi.ecs
+
+import java.net.{ InetAddress, NetworkInterface }
+
+import scala.collection.JavaConverters._
+
+import akka.annotation.ApiMayChange
+
+@ApiMayChange
+object AsyncEcsDiscovery {
+
+  // InetAddress.getLocalHost.getHostAddress throws an exception when running
+  // in awsvpc mode because the container name cannot be resolved.
+  // ECS provides a metadata file
+  // (https://docs.aws.amazon.com/AmazonECS/latest/developerguide/container-metadata.html)
+  // that we ought to be able to use instead to find our IP address, but the
+  // metadata file does not get set when running on Fargate. So this is our
+  // only option for determining what the canonical Akka and akka-management
+  // hostname values should be set to.
+  def getContainerAddress: Either[String, InetAddress] =
+    NetworkInterface.getNetworkInterfaces.asScala
+      .flatMap(_.getInetAddresses.asScala)
+      .filterNot(_.isLoopbackAddress)
+      .filter(_.isSiteLocalAddress)
+      .toList match {
+      case List(value) =>
+        Right(value)
+
+      case other =>
+        Left(s"Exactly one private address must be configured (found: $other).")
+    }
+}
