+++ integration-test/aws-api-ecs/src/main/scala/akka/cluster/bootstrap/EcsApiDemoApp.scala
@@ -7,7 +7,7 @@ package akka.cluster.bootstrap
 import java.net.InetAddress
 
 import akka.actor.ActorSystem
-import akka.discovery.awsapi.ecs.AsyncEcsServiceDiscovery
+import akka.discovery.awsapi.ecs.AsyncEcsDiscovery
 import akka.management.cluster.bootstrap.ClusterBootstrap
 import akka.management.scaladsl.AkkaManagement
 import com.typesafe.config.ConfigFactory
@@ -37,7 +37,7 @@ object EcsApiDemoApp {
   }
 
   private[this] def getPrivateAddressOrExit: InetAddress =
-    AsyncEcsServiceDiscovery.getContainerAddress match {
+    AsyncEcsDiscovery.getContainerAddress match {
       case Left(error) =>
         System.err.println(s"$error Halting.")
         sys.exit(1)
